+++ crates/graphql-composition/tests/composition_tests.rs
@@ -1,4 +1,4 @@
-use std::{fs, path::Path};
+use std::{fmt::Write as _, fs, path::Path};
 
 use graphql_composition::FederatedGraph;
 
@@ -47,33 +47,37 @@ fn run_test(test_path: &Path) -> anyhow::Result<()> {
             .map(|extension| graphql_composition::LoadedExtension::new(extension.url, extension.name)),
     );
 
-    let (federated_sdl, api_sdl) = match graphql_composition::compose(&subgraphs).into_result() {
-        Ok(federated_graph) => (
+    let result = graphql_composition::compose(&subgraphs);
+
+    let diagnostics = result.diagnostics();
+    let mut rendered_diagnostics = String::new();
+
+    for diagnostic in diagnostics.iter() {
+        let emoji = match diagnostic.severity() {
+            graphql_composition::diagnostics::Severity::Error => "❌",
+            graphql_composition::diagnostics::Severity::Warning => "⚠️",
+        };
+
+        let message = diagnostic.message();
+        writeln!(rendered_diagnostics, "- {emoji} {message}").unwrap();
+    }
+
+    let (federated_sdl, api_sdl) = if let Ok(federated_graph) = graphql_composition::compose(&subgraphs).into_result() {
+        (
             graphql_composition::render_federated_sdl(&federated_graph).expect("rendering federated SDL"),
-            Some(graphql_composition::render_api_sdl(&federated_graph)),
-        ),
-        Err(diagnostics) => (
-            format!(
-                "{}\n",
-                diagnostics
-                    .iter_messages()
-                    .map(|msg| format!("# {}", msg.lines().collect::<Vec<_>>().join("\\n")))
-                    .collect::<Vec<_>>()
-                    .join("\n"),
-            ),
-            None,
-        ),
+            graphql_composition::render_api_sdl(&federated_graph),
+        )
+    } else {
+        (String::new(), String::new())
     };
 
     let test_description = Some(test_description.as_str().trim())
         .filter(|desc| !desc.is_empty())
         .unwrap_or("Federated SDL");
 
+    insta::assert_snapshot!("diagnostics", rendered_diagnostics, test_description);
     insta::assert_snapshot!("federated.graphql", federated_sdl, test_description);
-
-    if let Some(actual_api_sdl) = api_sdl {
-        insta::assert_snapshot!("api.graphql", actual_api_sdl);
-    }
+    insta::assert_snapshot!("api.graphql", api_sdl, test_description);
 
     check_federated_sdl(&federated_sdl, test_path)
 }